// src/model/AccountQuota.ts
/**
 * @api.video/nodejs-client
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * The version of the OpenAPI document: 1
 *
 *
 * NOTE: This class is auto generated.
 * Do not edit the class manually.
 */

import AttributeType from './AttributeType';

/**
 * Deprecated
 */
export default class AccountQuota {
  /**
   * Deprecated
   */
  'quotaUsed'?: number;
  /**
   * Deprecated
   */
  'quotaRemaining'?: number;
  /**
   * Deprecated
   */
  'quotaTotal'?: number;

  static readonly discriminator?: string = undefined;

  static readonly attributeTypeMap: Array<AttributeType> = [
    {
      name: 'quotaUsed',
      baseName: 'quotaUsed',
      type: 'number',
      format: '',
    },
    {
      name: 'quotaRemaining',
      baseName: 'quotaRemaining',
      type: 'number',
      format: '',
    },
    {
      name: 'quotaTotal',
      baseName: 'quotaTotal',
      type: 'number',
      format: '',
    },
  ];

  static getAttributeTypeMap(): Array<AttributeType> {
    return AccountQuota.attributeTypeMap;
  }
}